Comment out exception in EntityAssociator.
There was a check if the associated entity is not null(1). But in the case when the product does not have default billing plan(which is valid situation) the check was not allowing to save the data. This edit can hide some bugs and errors in the code. So when the situation happens it has to be logged and the log has to be viewed by some developer.

(1) For example: if product#getDefaultBillingPlan() does not return null.